Fix MSVC warning C4804: unsafe use of type 'bool' in operation

+++ 3party/succinct/bit_vector.hpp
@@ -41,7 +41,7 @@ namespace succinct {
         bit_vector_builder(uint64_t size = 0, bool init = 0)
             : m_size(size)
         {
-            m_bits.resize(detail::words_for(size), uint64_t(-init));
+            m_bits.resize(detail::words_for(size), uint64_t(-static_cast<int64_t>(init)));
             if (size) {
                 m_cur_word = &m_bits.back();
                 // clear padding bits
